backend/requirements.txt
manim==0.17.3
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
streamlit==1.32.0
groq==0.4.1 